fix logic to actually stop at eof
Some encrypted files contain extra bits, so that aes has 16 bit to work on. in case of unknown n reading might fail (as the algorithm does not know it should stop until it bails. so either the padding has some realy interpretational value (contains 8 times 7 is that a hint?) or I need to check if the file conains enough values for a single row